feat(crosstarget-utils): introduce common trait "RegExprCompact" for "regex" util
