+++ libs/crosstarget-utils/src/common/timeout.rs
@@ -1,0 +1,12 @@
+use std::fmt::Display;
+
+#[derive(Debug)]
+pub struct TimeoutError;
+
+impl Display for TimeoutError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "Operation timed out")
+    }
+}
+
+impl std::error::Error for TimeoutError {}
