+++ libs/crosstarget-utils/src/lib.rs
@@ -9,4 +9,5 @@ mod native;
 #[cfg(not(target_arch = "wasm32"))]
 pub use crate::native::*;
 
-pub use common::SpawnError;
+pub use crate::common::regex::RegExpCompat;
+pub use crate::common::spawn::SpawnError;
