+++ libs/crosstarget-utils/src/wasm/regex.rs
@@ -1,7 +1,7 @@
 use enumflags2::BitFlags;
 use js_sys::RegExp as JSRegExp;
 
-use crate::common::{RegExpError, RegExpFlags};
+use crate::common::regex::{RegExpCompat, RegExpError, RegExpFlags};
 
 pub struct RegExp {
     inner: JSRegExp,
@@ -18,11 +18,10 @@ impl RegExp {
             inner: JSRegExp::new(pattern, &flags),
         })
     }
+}
 
-    /// Searches for the first match of this regex in the haystack given, and if found,
-    /// returns not only the overall match but also the matches of each capture group in the regex.
-    /// If no match is found, then None is returned.
-    pub fn captures(&self, message: &str) -> Option<Vec<String>> {
+impl RegExpCompat for RegExp {
+    fn captures(&self, message: &str) -> Option<Vec<String>> {
         let matches = self.inner.exec(message);
         matches.map(|matches| {
             let mut captures = Vec::new();
@@ -38,8 +37,7 @@ impl RegExp {
         })
     }
 
-    /// Tests if the regex matches the input string.
-    pub fn test(&self, input: &str) -> bool {
+    fn test(&self, input: &str) -> bool {
         self.inner.test(input)
     }
 }
