+++ libs/crosstarget-utils/src/wasm/spawn.rs
@@ -4,7 +4,7 @@ use futures::TryFutureExt;
 use tokio::sync::oneshot;
 use wasm_bindgen_futures::spawn_local;
 
-use crate::common::SpawnError;
+use crate::common::spawn::SpawnError;
 
 pub fn spawn_if_possible<F>(future: F) -> impl Future<Output = Result<F::Output, SpawnError>>
 where
